+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added `tags` and `customData` to the error sending methods.
 * Error sending methods can generate a stacktrace automatically.
 * Added `setUser()` method with `RaygunUserInfo`.
+* Added `setVersion()` method.
 
 ## 0.1.0